pyproject.toml
[project]
name = "ragline-query-api"
version = "1.0.0"
description = "Vector database query API for document retrieval"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.109.0,<0.110.0",
    "uvicorn[standard]>=0.27.0,<0.28.0",
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "llama-index-core>=0.10.20,<0.11.0",
    "llama-index-vector-stores-postgres>=0.1.6,<0.2.0",
    "llama-index-embeddings-bedrock>=0.1.2,<0.2.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "SQLAlchemy>=2.0.0,<3.0.0",
    "pgvector>=0.2.0,<0.3.0",
    "boto3>=1.34.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]
